wip: Trying to recreate the map via refactor
Map is currently not working as wanted:
- map marker is not refreshing everytime map is loaded
