+++ app/src/main/java/com/github/cstrerath/uncover/domain/map/models/MapBounds.kt
@@ -1,0 +1,8 @@
+package com.github.cstrerath.uncover.domain.map.models
+
+data class MapBounds(
+    val north: Double,
+    val east: Double,
+    val south: Double,
+    val west: Double
+)
