+++ app/src/main/java/com/github/cstrerath/uncover/ui/navigation/MenuItems.kt
@@ -11,7 +11,6 @@ enum class MenuItems(
     val action: (NavigationManager) -> Unit
 ) {
     MAP(R.string.menu_map, { it.navigateToMap() }),
-    CHARACTER_LIST(R.string.menu_character_list, { it.navigateToCharacterList() }),
     CHARACTER_CREATION(R.string.menu_character_creation, { it.navigateToCharacterCreation() }),
     LOCATIONS(R.string.menu_locations, { it.navigateToLocations() }),
     PLAYER_STATS(R.string.menu_player_stats, { it.navigateToPlayerStats() })
